+++ cloud/acp-admin-cloud-constant/src/main/kotlin/pers/acp/admin/constant/RouteConstant.kt
@@ -5,13 +5,15 @@ package pers.acp.admin.constant
  * @since JDK 11
  */
 object RouteConstant {
-    const val UPDATE_ROUTE_INPUT = "updateRouteConsumer"
-    const val UPDATE_GATEWAY_OUTPUT = "updateRouteProducer"
+    const val UPDATE_ROUTE_CONSUMER = "updateRouteConsumer"
+    const val UPDATE_ROUTE_INPUT = "updateRouteInput"
+    const val UPDATE_ROUTE_OUTPUT = "updateRouteOutput"
     const val UPDATE_ROUTE_DESCRIPTION = "update-gateway-routes"
     const val ROUTES_DEFINITION_KEY = "acp_cloud_gateway_routes_definition_key"
     const val UPDATE_GATEWAY_ROUTES = "update_gateway_routes"
-    const val ROUTE_LOG_INPUT = "routeLogConsumer"
-    const val ROUTE_LOG_OUTPUT = "routeLogProducer"
+    const val ROUTE_LOG_CONSUMER = "routeLogConsumer"
+    const val ROUTE_LOG_INPUT = "routeLogInput"
+    const val ROUTE_LOG_OUTPUT = "routeLogOutput"
     const val ROUTE_LOG_DESCRIPTION = "route-log"
     const val ROUTE_LOG_CONSUMER_GROUP = "route-log-group"
 }